Define dummy variables, exclude outliers
Plots of target variable against a single numerical predictor are being
used to find cases where feature engineering would improve a multilinear
model.  In some of those plots, the fit is visibly distorted by outliers
and/or by the points at the intercept.  Exclude outliers and define
dummy variables to allow nonzero values of individual predictors to be
fit independently.